api: finalised enhacements to reusability by separating functions by usage
